+++ src/lib/mina/smart-contract-context.ts
@@ -2,7 +2,7 @@ import type { SmartContract } from '../zkapp.js';
 import type {
   AccountUpdate,
   UnfinishedForest,
-  UnfinishedTree,
+  AccountUpdateLayout,
 } from '../account_update.js';
 import { Context } from '../global-context.js';
 
@@ -17,33 +17,3 @@ type SmartContractContext = {
 let smartContractContext = Context.create<null | SmartContractContext>({
   default: null,
 });
-
-const SmartContractContext = {
-  enter(self: SmartContract, selfUpdate: AccountUpdate) {
-    let context: SmartContractContext = {
-      this: self,
-      selfUpdate,
-      selfLayout: new AccountUpdateLayout(),
-      selfCalls: { useHash: false, value: [] },
-    };
-    let id = smartContractContext.enter(context);
-    return { id, context };
-  },
-  leave(id: number) {
-    smartContractContext.leave(id);
-  },
-  stepOutside() {
-    return smartContractContext.enter(null);
-  },
-  get() {
-    return smartContractContext.get();
-  },
-};
-
-class AccountUpdateLayout {
-  map: Map<bigint, UnfinishedTree>;
-
-  constructor() {
-    this.map = new Map();
-  }
-}
